Returned decimal number for price.

+++ library/src/main/java/ua/syi/billing/library/models/PlayMarketConsumable.kt
@@ -5,7 +5,7 @@ import com.android.billingclient.api.ProductDetails
 class PlayMarketConsumable(
     val id: String,
     val formattedPrice: String,
-    val price: Long,
+    val price: Double,
     val currency: String,
     internal val payload: ProductDetails,
 ) : Purchasable {
